+++ resources/usr/sbin/setupMainNode.sh
@@ -3,11 +3,15 @@ set -o errexit
 set -o nounset
 set -o pipefail
 
+source k3s-util.sh
+
 nodeIp=${1}
 nodeExternalIp=${2}
 flannelInterface=${3}
 k3sToken=${4}
 username=${5}
+nodeLabels="${6}"
+nodeTaints="${7}"
 
 k3sVersion=$(cat /var/lib/rancher/k3s/agent/images/k3sVersion)
 
@@ -20,13 +24,18 @@ else
   systemctl restart systemd-timesyncd.service
 fi
 
+nodeLabelOptions=$(getOptionListForList "--node-label" "${nodeLabels}")
+nodeTaintOptions=$(getOptionListForList "--node-taint" "${nodeTaints}")
+
 echo "Installing k3s ${k3sVersion}..."
 INSTALL_K3S_SKIP_DOWNLOAD=true \
 INSTALL_K3S_VERSION=${k3sVersion} \
 K3S_KUBECONFIG_MODE="644" \
 K3S_TOKEN=${k3sToken} \
 K3S_EXTERNAL_IP="${nodeExternalIp}" \
-INSTALL_K3S_EXEC="--disable local-storage
+INSTALL_K3S_EXEC="${nodeLabelOptions}
+ ${nodeTaintOptions}
+ --disable local-storage
  --node-label svccontroller.k3s.cattle.io/enablelb=true
  --disable traefik
  --node-external-ip=${nodeExternalIp}