+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade k3s to 1.28.3 #56
 - Passwords (Docker-, Dogu- & Helmregistry) has to be encoded in Base64 (see [here](docs/development/dev_box_en.md) and [here](terraform/ces-module/README.md)) #64
 - Disk space related optimizations for development (f.e. fewer longhorn replicas) #71
+- Add options to add node labels and taints on cluster setup #73
 
 ### Added
 - Packer templates for CES production images